Moving around + fixes
