+++ tests/test_mc.py
@@ -42,7 +42,7 @@ class LogFunc(Func):
 
 
 def test_reinforce_estimator(device, is_log):
-    N, T = int(1e6), 30
+    N, T = int(1e5), 30
     logits = torch.randn(T, device=device, requires_grad=True)
     mask = torch.randint(2, (T,), device=device) == 1
     probs = logits.sigmoid().masked_fill(mask, 0)
@@ -118,7 +118,8 @@ def test_importance_sampling_estimator(device, self_normalize, is_log):
         assert (g_q_probs == 0).all()
 
 
-def test_rebar_estimator_bernoulli(device, is_log):
+@pytest.mark.parametrize("varmin", [True, False], ids=["varmin", "mc"])
+def test_rebar_estimator_bernoulli(device, is_log, varmin):
     N, T = int(1e5), 30
     logits = torch.randn(T, device=device, requires_grad=True)
     mask = torch.randint(2, (T,), device=device) == 1
@@ -132,9 +133,11 @@ def test_rebar_estimator_bernoulli(device, is_log):
     probs = logits.sigmoid().masked_fill(mask, 0)
     dist = distributions.LogisticBernoulli(probs=probs)
     cv = modules.LogisticBernoulliRebarControlVariate(func).to(device)
-    estimator = estimators.RelaxEstimator(
-        dist, func, cv, [probs], [cv.log_temp, cv.eta], is_log=is_log
-    )
+    if varmin:
+        args = [probs], [cv.log_temp, cv.eta]
+    else:
+        args = tuple()
+    estimator = estimators.RelaxEstimator(dist, func, cv, *args, is_log=is_log)
     v = estimator.estimate(N).sum()
     act_loss = (v - T / 2) ** 2
     assert torch.isclose(exp_loss, act_loss, atol=1)
@@ -148,7 +151,8 @@ def test_rebar_estimator_bernoulli(device, is_log):
     assert not torch.isclose(g_eta, zero_)
 
 
-def test_rebar_estimator_categorical(device):
+@pytest.mark.parametrize("varmin", [True, False], ids=["varmin", "mc"])
+def test_rebar_estimator_categorical(device, varmin):
     N, T, V = int(1e5), 12, 5
     logits = torch.randn((T, V), device=device, requires_grad=True)
     mask = torch.randint_like(logits, 2) == 1
@@ -168,9 +172,11 @@ def test_rebar_estimator_categorical(device):
     probs = logits_.softmax(-1)
     dist = distributions.GumbelOneHotCategorical(probs=probs)
     cv = modules.GumbelOneHotCategoricalRebarControlVariate(func).to(device)
-    estimator = estimators.RelaxEstimator(
-        dist, func, cv, [probs], [cv.log_temp, cv.eta]
-    )
+    if varmin:
+        args = [probs], [cv.log_temp, cv.eta]
+    else:
+        args = tuple()
+    estimator = estimators.RelaxEstimator(dist, func, cv, *args)
     v = estimator.estimate(N).sum()
     act_loss = (v - T / 2) ** 2
     assert torch.isclose(exp_loss, act_loss, atol=1)
